requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pillow==12.0.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
WTForms==3.2.1
